Lift form state to app level
